sigma_tool/backend/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==22.0
pymongo==4.3.3
pyparsing==3.0.9
pysigma==0.9.4
pysigma-backend-splunk==1.0.2
PyYAML==6.0
requests==2.28.2
six==1.16.0
smmap==5.0.0
urllib3==1.26.14
Werkzeug==2.2.3
